ni_spec_server_proxy/pyproject.toml
[tool.poetry]
name = "ni_spec_server_proxy"
version = "1.2.0"
description = "Python package to redirect SCM API calls to SLE APIs to exchange data with SLE Server."
authors = ["NI <opensource@ni.com>"]
repository = "https://github.com/ni/ni-spec-server-proxy"
license = "MIT"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: Flask",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: Microsoft :: Windows",
    # Poetry automatically adds classifiers for the license and the supported Python versions.
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering",
]
readme = "README.md"
packages = [{ include = "ni_spec_server_proxy", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^3.0.3"
asgiref = "^3.8.1"
nisystemlink-clients = "^1.3.0"
systemlink-sdk = "^24.0.0"
python-dateutil = "^2.8.2"
aiohttp = "^3.8.5"
backoff = "^2.2.1"
six = "^1.15.0"
certifi = "^2024.0.0"
pyyaml = "^6.0.0"
pandas = "1.0.5"
numpy = "1.19.0"

[[tool.poetry.source]]
name = "NI PyPI"
url = "https://pypi.ni.systems"
default = true

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
ni-python-styleguide = "^0.4.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
